Move asset related code to its own module
