+++ albam/blender_ui/__init__.py
@@ -1,3 +1,7 @@
+from .asset import (
+    ALBAM_PT_AssetObject,
+    ALBAM_PT_AssetImage,
+)
 from .custom_properties import (
     ALBAM_PT_CustomPropertiesMaterial,
     ALBAM_PT_CustomPropertiesMesh,
@@ -6,6 +10,9 @@ from .error_handling import ALBAM_OT_ErrorHandler
 
 
 __all__ = (
+    "ALBAM_PT_AssetImage",
+    "ALBAM_PT_AssetObject",
+    "ALBAM_PT_CustomPropertiesMesh",
     "ALBAM_PT_CustomPropertiesMaterial",
     "ALBAM_PT_CustomPropertiesMesh",
     "ALBAM_OT_ErrorHandler",
